Give the transcription configuration its own file
