+++ Sources/Nativ/Features/VoiceCapture/VoiceCaptureCoordinator.swift
@@ -1,16 +1,6 @@
 import AppKit
 import NativServerKit
 
-struct VoiceTranscriptionConfiguration: Sendable {
-    let modelSearchPath: String
-    let additionalModelSearchPaths: [String]
-    let selectedModelID: String?
-    let languageModelID: String?
-    let maxTokens: Int
-    let serverBaseURL: URL
-    let serverAPIKey: String?
-    let serverIsRunning: Bool
-}
 
 @MainActor
 final class VoiceCaptureCoordinator {
